fix:[sc-19] replace nodemailer configuration from utils to helper
