--- src/utils/mail-sender.js
@@ -1,29 +1,0 @@
-const nodemailer = require('nodemailer');
-
-const mailSender = async (email, title, body) => {
-  const transporter = nodemailer.createTransport({
-    host: process.env.MAIL_HOST,
-    port: process.env.MAIL_PORT,
-    secure: true,
-    auth: {
-      user: process.env.MAIL_USER,
-      pass: process.env.MAIL_PASS,
-    },
-  });
-
-  // console.log('Created transporter: ');
-  const mailOptions = {
-    from: process.env.EMAIL || 'Jyoti Deveda',
-    to: email,
-    subject: title,
-    text: body,
-  };
-
-  const info = await transporter.sendMail(mailOptions);
-  // console.log('Mail info: ', info);
-  return info;
-};
-
-module.exports = {
-  mailSender,
-};
